refactor: address react-doctor findings on commentary feature
- Extract all modal state/effects/handlers into a use-commentary hook
  and split the view into ScopeBar + CommentaryOutput (no-giant-component)
- Remount the modal per open via a React key instead of resetting state
  when isOpen/initialSelection change (no-adjust-state-on-prop-change)
- Collect assistant text in one pass instead of a chained
  filter/flatMap/map (js-combine-iterations)
